+++ CHANGELOG.md
@@ -1,3 +1,11 @@
+v0.0.19
+---
+- Removed unused parameter
+ 
+v0.0.18
+---
+- Added backend and frontend port flags to up command
+ 
 v0.0.17
 ---
 - Updating sample ECS dashboard with parameter support